recommendation/requirements.txt
requests==2.21.0
Flask==1.0.2
Flask_Cors==3.0.8
python-dotenv==0.12.0